requirements.txt
fastapi
pydantic
uvicorn
langchain
langchain-google-genai
sentence-transformers
faiss-cpu
python-dotenv
langchain-community